requirements.txt
fastapi
uvicorn[standard]
azure-devops>=7.1.0b4
msrest>=0.7.1
python-dotenv>=1.0.0
